Added Date and  modified sub-optimal transfer
Added date in the plot

Added new part for computation of V infinity nearest the assigned one to
plot the sub-optimal transfer arc.
The assigned launcher have a V_inf of 5.8 Km/s  so we have to use the
nearest value that we find (is about 5,7979) for our suboptimal
transfer, that is really close the Dv_min trasnfer.